+++ .circleci/install-system.sh
@@ -4,6 +4,7 @@ set -eux
 apt-get update
 apt-get install -y \
   emacs \
+  git \
   libcurl4-openssl-dev \
   libssl-dev \
   libxml2-dev \