+++ R/zzz.R
@@ -8,6 +8,20 @@
   packageStartupMessage(paste(m, collapse = "\n"))
 }
 
+.onLoad <- function(libname, pkgname) {
+  sysgit <- Sys.which("git")
+  wflow_pkg_opts <- list(
+    workflowr.sysgit = if(fs::file_exists(sysgit)) sysgit else "",
+    workflowr.view = TRUE
+  )
+
+  op <- options()
+  toset <- !(names(wflow_pkg_opts) %in% names(op))
+  if(any(toset)) options(wflow_pkg_opts[toset])
+
+  invisible()
+}
+
 #' workflowr: A workflow template for creating a research website
 #'
 #' The workflowr package helps you create a research website using R Markdown
@@ -48,8 +62,29 @@
 #'   project-wide rendering settings can be customized in the
 #'   \code{_site.yml} file.}
 #' }
-#' 
+#'
+#' @section Package options:
+#'
+#' The following package options affect the default behavior of the workflowr
+#' functions. To permanently set any of these options, add a call to the
+#' function \code{\link[base]{options}} in the file \code{.Rprofile} at the root
+#' of your workflowr project. For example:
+#'
+#' \preformatted{
+#' # Do not use Git executable
+#' options(workflowr.sysgit = "")
+#' }
+#'
+#' \describe{
+#'
+#' \item{workflowr.sysgit}{The path to the system Git executable. This is
+#' occasionally used to increase the speed of Git operations performed by
+#' workflowr. By default it is set to the first Git executable on the search
+#' path. You can specify a path to a different Git executable. Alternatively you
+#' can disable this behavior entirely by setting it to the empty string \code{""}.}
+#' }
+#'
 #' @docType package
 #' @name workflowr
-#' 
+#'
 NULL